remove duplicated geometric types, fixes filling ellipses

+++ libnanocv/geom.h
@@ -7,7 +7,7 @@ namespace ncv
         ///
         /// \brief image coordinate
         ///
-        typedef int16_t                 coord_t;
+        typedef int32_t                 coord_t;
         typedef int32_t                 area_t;
 }
